requirements.txt
pypdf
PyMuPDF
Pillow
